requirements.txt
# AI APIs
anthropic>=0.45.0  # Latest Anthropic SDK with Sonnet 4.5 support

# Web Framework
flask>=3.1.0
python-dotenv>=1.0.0

# HTTP Client
requests>=2.32.0

# Future integrations (uncomment when needed)
# google-generativeai>=0.8.0  # Gemini API
# openai>=1.0.0  # ChatGPT API
# notion-client>=2.0.0  # Notion API
# slack-sdk>=3.0.0  # Slack API
